refactor: extract in-process subagent tools
